20191028 update account private member type

+++ src/model/CurrentAccount.java
@@ -1,8 +1,10 @@
 package model;
 
+import java.math.BigDecimal;
+
 public class CurrentAccount extends Account {
 
-    public CurrentAccount(String id, double balance, String customerName) {
+    public CurrentAccount(String id, BigDecimal balance, String customerName) {
         super(id, balance, customerName);
     }
 }
